Analytics.Datalog.Internal

Description

This module is not considered packaged under the package versioning policy. Any direct dependency upon it is likely to break even between minor versions.

Documentation

class Traversable t => Match t where

Simple flat unification

Methods

match :: (a -> b -> c) -> t a -> t b -> Maybe (t c)

Use Generic1 by default

default match :: (Generic1 t, GMatch (Rep1 t)) => (a -> b -> c) -> t a -> t b -> Maybe (t c)

class (Match v, Eq1 v) => Variable v where

A Variable type has a distinguishable var construction.

Methods

var :: Prism (v a) (v b) a b

matchVar :: Variable t => (a -> b -> c) -> t a -> t b -> Maybe (t c)

Default definition of match for a Var

Datalog

data Datalog where

An operational encoding of a datalog program.

Constructors

Fact :: (Typeable1 t, Match t) => (forall a. t a) -> Datalog m ()

:- :: Ord a => Relation a -> Body a t -> Datalog m ()

Query :: Ord a => Body a t -> Datalog m [t]

Bind :: Datalog m a -> (a -> Datalog m b) -> Datalog m b

Return :: a -> Datalog m a

Lift :: m a -> Datalog m a

query :: Ord a => Body a t -> Datalog m [t]

data Relation v

A relation with free variables of type a. This is used to represent the head of a datalog rule and is used by no.

Constructors

forall t . (Typeable1 t, Match t) => Relation (t v)

data Body v a where

This models a strongly typed query or the right hand side of a Datalog rule.

The Body itself forms an Alternative, letting you combine them to make a robust query language.

Constructors

Ap :: Body v (a -> b) -> Body v a -> Body v b

Map :: (a -> b) -> Body v a -> Body v b

Pure :: a -> Body v a

Alt :: Body v a -> Body v a -> Body v a

Empty :: Body v a

Body :: (Typeable1 t, Match t) => t v -> Body v (t a)

No :: Relation v -> Body v ()

no :: Relation a -> Body a ()

Stratified negation.

Implementation Details

class (Typeable1 t, Match t) => Rel t a r | r -> a where

This provides overloading of predicates so they can be used directly as facts in the Datalog Monad or as the head or Body of a rule or as a query.

Methods

rel :: t a -> r