Data.Analytics.Active

Task

data Task a

Constructors

Task

Fields

runTask :: (Job -> ContT () IO ()) -> ContT () IO a

class MonadIO m => MonadTask m where

Methods

spawn :: Task () -> m ()

run :: MonadIO m => Task a -> m a

Eventually we can do work stealing. For now we just push work onto the channel and keep taking the next item. Other workers can take from the job pool too

(|>>) :: Task a -> Task b -> Task b

Spawn a background task and continue

Observer

data Observer a

Constructors

Observer

Fields

(!) :: a -> Task ()

kill :: SomeException -> Task ()

complete :: Task ()

foreach :: (a -> Task b) -> Observer a

Perform an action for each

>>> run $  subscribe (observe [1..10]) $ foreach (liftIO . print)

Observable

newtype Observable a

Constructors

Observable

Fields

subscribe :: Observer a -> Task Subscription

safe :: Observable a -> Observable a

Enforce the Observable protocol.

fby :: a -> Observable a -> Observable a

never :: Observable a

observe :: Foldable f => f a -> Observable a

Subscription

newtype Subscription

Constructors

Subscription

Fields

cancel :: Task ()