Data.Analytics.Active.Task

Documentation

data Task a

Constructors

Task

Fields

runTask :: (Job -> ContT () IO ()) -> ContT () IO a

class MonadIO m => MonadTask m where

Methods

spawn :: Task () -> m ()

run :: MonadIO m => Task a -> m a

Eventually we can do work stealing. For now we just push work onto the channel and keep taking the next item. Other workers can take from the job pool too

(|>>) :: Task a -> Task b -> Task b

Spawn a background task and continue

type Job = Task ()