Data.Analytics.Approximate.Log

Documentation

newtype Log a

data Log has no connection to Datalog.

Errors should be resolved by taking the answer to negative infinity where possible as this is typically used as the lower bound on a confidence level.

Constructors

Log a

class Floating a => Precise a where

Methods

log1p :: a -> a

expm1 :: a -> a