Data.Analytics.Approximate.Summation

Description

Monoidal compensated summation based on Knuth's error free transformation and various algorithms by Ogita, Rump and Oishi.

Documentation

data EFT

Constructors

EFT !Double !Double

runEFT :: EFT -> Double

magic :: Double

split :: Double -> EFT

error-free split of a floating point number into two parts.

sum2 :: Double -> Double -> EFT

Knuth's error free transformation

After EFT x y = sum2 a b:

 a + b = x + y
 x = fl(a + b)

sum3 :: Double -> Double -> Double -> EFT

Ogita, Rump and Oishi's VecSum

times2 :: Double -> Double -> EFT