Data.Analytics.Approximate.Type

Description

Documentation

data Approximate a

An approximate number, with a likely interval, an expected value and a lower bound on the log of probability that the answer falls in the interval.

NB: The probabilities associated with confidence are stored in the log domain.

Constructors

Approximate !(Log Double) a a a

class HasApproximate t a | t -> a where

Methods

approximate :: Lens' t (Approximate a)

confidence :: Lens' t (Log Double)

estimate :: Lens' t a

hi :: Lens' t a

lo :: Lens' t a

exact :: Eq a => Prism' (Approximate a) a

zero :: (Num a, Eq a) => Prism' (Approximate a) ()

one :: (Num a, Eq a) => Prism' (Approximate a) ()

withMin :: Ord a => a -> Approximate a -> Approximate a

withMax :: Ord a => a -> Approximate a -> Approximate a