Data.Analytics.Bits

Description

Documentation

class (Num t, Bits t) => Ranked t where

Methods

lsb :: t -> Int

Calculate the least significant set bit using a debruijn multiplication table. NB: The result of this function is undefined when given 0.

rank :: t -> Int

Calculate the number of trailing 0 bits.

log2 :: Word32 -> Int

w32 :: Integral a => a -> Word32

w64 :: Integral a => a -> Word64