Data.Analytics.Compression.Arithmetic

Description

http://www.stanford.edu/class/ee398a/handouts/papers/WittenACM87ArithmCoding.pdf http://www.sable.mcgill.ca/publications/techreports/2007-5/bodden-07-arithmetic-TR.pdf

Encoding

data Encoding

encoding :: MonadPut m => Encoding -> Coding m ()

Determine any final trailing e3 bits.

encode :: MonadPut m => Int64 -> Int64 -> Int64 -> Encoding -> Coding m Encoding

encode lo hi total will encode a symbol in our arithmetic encoding with weight (hi - lo) >= 1 out of total probability.

Decoding

newtype Decoding m a

Constructors

Decoding

Fields

runDecoding :: forall r. (a -> Int -> Int64 -> Coding m r) -> Int -> Int64 -> Coding m r

decoding :: MonadGet m => Decoding m a -> Coding m a

For now fill all but the most significant bit directly.

This means we cannot arithmetic encode a sequence of bits shorter than 63 bits.