Data.Analytics.Datalog.Prompt

Description

This module lets you step through a Datalog program one rule or Query at a time.

Note: This module conflicts with Data.Analytics.Datalog.Monad and the API is mostly of interest to query engine providers, so it is not re-exported through Data.Analytics.Datalog.

Each Step

type Step = StepT Identity

data StepT where

A single Datalog rule or Query.

Constructors

Fresh :: (a -> a -> a) -> StepT m (Table a)

:- :: Atom a b -> Query a -> StepT m ()

Query :: Query a -> StepT m (Logic a)

Prompting for Steps

type Prompt = PromptT Identity

data PromptT where

The result of prompting, after monadic effects.

Constructors

Done :: a -> PromptT m a

:>>= :: StepT m a -> (a -> DatalogT m b) -> PromptT m b

prompt :: Datalog a -> Prompt a

promptT :: Monad m => DatalogT m a -> m (PromptT m a)

Quotient out operational details of the Datalog program and get to the next Step.

unprompt :: PromptT m a -> DatalogT m a

unpromptT :: m (PromptT m a) -> DatalogT m a

Prompt isomorphisms

_Prompt :: Iso (Datalog a) (Datalog b) (Prompt a) (Prompt b)

_PromptT :: Monad m => Iso (DatalogT m a) (DatalogT n b) (m (PromptT m a)) (n (PromptT n b))