Data.Analytics.Datalog.Query

Description

Documentation

data Query where

This models a strongly typed query or the right hand side of a Datalog rule.

The Query itself forms an Alternative, letting you combine them to make a robust query language.

Constructors

Ap :: Query (a -> b) -> Query a -> Query b

Map :: (a -> b) -> Query a -> Query b

Pure :: a -> Query a

Alt :: Query a -> Query a -> Query a

Empty :: Query a

Value :: (Typeable a, Typeable b) => Atom a b -> Query a

Row :: (Typeable a, Typeable b) => Atom a b -> Query b

Key :: Term a => a -> Query (Entity a)

No :: (Typeable a, Typeable b) => Atom a b -> Query ()

no :: (Typeable a, Typeable b) => Atom a b -> Query ()

Stratified negation.

key :: Term a => a -> Query (Entity a)

row :: (Typeable a, Typeable b) => Atom a b -> Query b

value :: (Typeable a, Typeable b) => Atom a b -> Query a