Data.Analytics.File

Description

Documentation

data InsufficientDiskSpace

class AsInsufficientDiskSpace t where

Methods

_InsufficientDiskSpace :: Prism' t InsufficientDiskSpace

allocate :: Handle -> Int64 -> IO ()

Expand a file with contiguous disk storage where possible.

prefetch :: Handle -> Int64 -> Int64 -> IO ()

Prefetch a range from a file on disk.

prefetch h o c signals to the operating system that we intend to access the bytes at positions [o..o+c) in the file with handle h soon.

sync :: Handle -> IO ()

Sync a file to disk.