Data.Analytics.Internal.Datalog

Description

This module is not considered packaged under the package versioning policy. Any direct dependency upon it is likely to break even between minor versions.

Datalog

data Datalog where

An operational encoding of a Datalog program.

Constructors

Fact :: (Typeable1 t, Match t) => (forall a. t a) -> Datalog m ()

:- :: Ord a => Relation a -> Query a t -> Datalog m ()

Query :: Ord a => Query a t -> Datalog m [t]

Bind :: Datalog m a -> (a -> Datalog m b) -> Datalog m b

Return :: a -> Datalog m a

Lift :: m a -> Datalog m a

query :: Ord a => Query a t -> Datalog m [t]