Data.Analytics.Key

Description

Documentation

data Interleaving a

This provides a priority queue with a fixed schedule that can be used to obtain an interleaved sequence of bits from multiple sources. Use the monoid to fairly interleave breaking ties to the left.

Constructors

Z

Interleaving !(Approximate Int) !Int !(Heap a)

interleaving :: Schedule a -> a -> Interleaving Bool

interleavingBy :: Schedule a -> [b] -> Interleaving b

Schedules

data Schedule a

This is used to schedule how we interleave our keys.

class HasSchedule t a | t -> a where

Methods

schedule :: Lens' t (Schedule a)

scheduleApproximate :: Lens' t Bool

scheduleBits :: Lens' t (Approximate Int)

scheduleEncoder :: Lens' t (a -> [Bool])

scheduleOrdered :: Lens' t Bool

schedulePriority :: Lens' t Int

scheduleStride :: Lens' t Int

class Scheduled a where

Retrieve the default key schedule for a type

Methods

scheduled :: Schedule a

integral :: forall a. (Storable a, Bits a) => Schedule a

bits :: Bits a => Int -> Schedule a

hashed :: Hashable a => Schedule a

Character schedules

ascii :: Schedule Char

iso8859_1 :: Schedule Char

We conservatively approximate characters above of this range by the upper bound to preserve ordering.