Data.Analytics.Key.Interleaving

Description

This provides a generalized key interleaving system for variable-length keys. For fixed-length keys this looks like the Morton or Z-order.

TODO: Provide for Grey-coding for (fixed length parts of?) the resulting schedule to get us a variable-length Hilbert ordering.

Documentation

data Interleaving a

This provides a priority queue with a fixed schedule that can be used to obtain an interleaved sequence of bits from multiple sources. Use the monoid to fairly interleave breaking ties to the left.

Constructors

Z

Interleaving !(Approximate Int) !Int !(Heap a)

interleaving :: Schedule a -> a -> Interleaving Bool

interleavingBy :: Schedule a -> [b] -> Interleaving b