Data.Analytics.Morton.Schedule

Description

Documentation

data Schedule a

This is used to schedule how we interleave our keys.

Constructors

Schedule !Int !Int !(Approximate Int) !Bool !Bool (a -> [Bool])

class HasSchedule t a | t -> a where

Methods

schedule :: Lens' t (Schedule a)

scheduleApproximate :: Lens' t Bool

scheduleBits :: Lens' t (Approximate Int)

scheduleEncoder :: Lens' t (a -> [Bool])

scheduleOrdered :: Lens' t Bool

schedulePriority :: Lens' t Int

scheduleStride :: Lens' t Int

class Scheduled a where

Retrieve the default key schedule for a type

Methods

scheduled :: Schedule a

magic :: Int

>>> foldr lcm 2 [3,4,5,6,7,8,9,10,11,12,13,14,15,20,23,31,63]
256936680

integral :: forall a. (Storable a, Bits a) => Schedule a

bits :: Bits a => Int -> Schedule a

hashed :: Hashable a => Schedule a

Character schedules

ascii :: Schedule Char

iso8859_1 :: Schedule Char

We conservatively approximate characters above of this range by the upper bound to preserve ordering.