Data.Analytics.Morton.Type

Documentation

data Morton f

This provides a double-ended priority queue with a fixed schedule that can be used to obtain a generalized morton sequence of bits from multiple sources. Use the monoid to fairly interleave breaking ties to the left.

Constructors

Z

Morton !Int !Int !(MinHeap (f (Int -> Bool))) !(MaxHeap (f (Int -> Bool)))

morton :: Functor f => Schedule a -> f a -> Morton f

morton64 :: Functor f => Schedule a -> f (Int -> Bool) -> Morton f