Data.Analytics.Prompt

Description

This module lets you step through a Datalog program one Fact, rule or Query at a time.

Note: This module conflicts with Data.Analytics.Datalog and Data.Analytics.Internal.Datalog and the API is mostly of interest to query engine providers, so it is not re-exported through Data.Analytics.

Documentation

data Step where

A single Datalog Fact, rule or Query.

Constructors

Fact :: (Typeable1 t, Match t) => (forall v. t v) -> Step m ()

:- :: Ord v => Relation v -> Query v t -> Step m ()

Query :: Ord v => Query v t -> Step m [t]

data Prompt where

The result of prompting, after monadic effects.

Constructors

Done :: a -> Prompt m a

:>>= :: Step m a -> (a -> Datalog m b) -> Prompt m b

prompt :: Monad m => Datalog m a -> m (Prompt m a)

Quotient out operational details of the Datalog program and get to the next Step.