Data.Analytics.Query

Description

Datalog

data Query v a

This models a strongly typed query or the right hand side of a Datalog rule.

The Query itself forms an Alternative, letting you combine them to make a robust query language.

no :: Relation a -> Query a ()

Stratified negation.