Data.Analytics.Relation

Description

Documentation

data Relation v

A relation with free variables of type a. This is used to represent the head of a datalog rule and is used by no.

Constructors

forall t . (Typeable1 t, Match t) => Relation (t v)

class (Typeable1 t, Match t) => Rel t a r | r -> a where

This provides overloading of predicates so they can be used directly as facts in the Datalog Monad or as the head or body of a rule or as a Query.

Methods

rel :: t a -> r