Data.Analytics.Statistics.Moments

Documentation

data Moments

The first few central moments.

Constructors

Moments

Fields

_count :: !Int64

_mean :: !Double

_moment2 :: !Double

_moment3 :: !Double

_moment4 :: !Double

class HasMoments t where

Methods

moments :: Lens' t Moments

count :: Lens' t Int64

mean :: Lens' t Double

moment2 :: Lens' t Double

moment3 :: Lens' t Double

moment4 :: Lens' t Double

variance :: HasMoments t => Traversal' t Double

Extract the variance as a moment around the mean.

NB: setting this will render the higher Moments inconsistent.

stddev :: HasMoments t => Traversal' t Double

NB: setting this will render the higher Moments inconsistent.

skewness :: HasMoments t => Fold t Double

Calculate skewness. This is a Fold because skewness may not be defined for some combinations of Moments.

kurtosis :: HasMoments t => Fold t Double

Calculate kurtosis. This is a Fold because kurtosis may not be defined for some combinations of Moments.

singleton :: Real a => a -> Moments

momentsOf :: Real a => Getting Moments s t a b -> s -> Moments