Data.Analytics.Statistics.Variances

Documentation

data Variances

Constructors

Variances !Int64 !Double !Double !Double !Double !Double

class HasVariances t where

Methods

variances :: Lens' t Variances

covXY :: Lens' t Double

meanX :: Lens' t Double

meanY :: Lens' t Double

moment2X :: Lens' t Double

moment2Y :: Lens' t Double

points :: Lens' t Int64

doubleton :: (Real a, Real b) => a -> b -> Variances

variancesOf :: (Real a, Real b) => Getting Variances s t (a, b) (a, b) -> s -> Variances

covariance :: HasVariances t => Traversal' t Double

correlation :: HasVariances t => Traversal' t Double

Extract the correlation coefficient

varianceX :: HasVariances t => Traversal' t Double

Extract the variance as a moment around the mean.

varianceY :: HasVariances t => Traversal' t Double

Extract the variance as a moment around the mean.

stddevX :: HasVariances t => Traversal' t Double

stddevY :: HasVariances t => Traversal' t Double