Data.Analytics.Succinct.Rank.Class

Description

Documentation

class Rank a t | t -> a where

Methods

rank :: a -> Int -> t -> Int

rank x i xs computes the number of occurrences of x in xs in positions [0..i)

A default definition is supplied in terms of Foldable.