Data.Analytics.Succinct.Rank.Pop

Description

A simple succinct Rank structure that uses popCount rather than a second level of indirection.

http://en.wikipedia.org/wiki/Succinct_data_structure

Documentation

data Pop

A very simple popCount-based succinct Rank structure.

Constructors

Pop

Fields

_popBits :: !(Vector Word64)

_popCounts :: !(Vector Int)

class HasPop t where

Methods

pop :: Lens' t Pop

popBits :: Lens' t (Vector Word64)

popCounts :: Lens' t (Vector Int)

new :: Vector Word64 -> Pop

Create a vector of bits enumerated from least significant bit of the 0th slot upward including a Rank structureazzza

rank :: Rank a t => a -> Int -> t -> Int

rank x i xs computes the number of occurrences of x in xs in positions [0..i)

A default definition is supplied in terms of Foldable.