Data.Analytics.Internal.Datalog

Description

This module is not considered packaged under the package versioning policy. Any direct dependency upon it is likely to break even between minor versions.

Datalog

type Datalog t = DatalogT t Identity

An operational encoding of a Datalog program.

data DatalogT where

An operational encoding of a Datalog program with extra effects in m.

Constructors

Fact :: Atom t a -> DatalogT t m ()

:- :: Atom t a -> Query t b -> DatalogT t m ()

Query :: Query t a -> DatalogT t m [a]

Bind :: DatalogT t m a -> (a -> DatalogT t m b) -> DatalogT t m b

Return :: a -> DatalogT t m a

Lift :: m a -> DatalogT t m a

query :: Query t a -> DatalogT t m [a]

Perform a Query.