Data.Analytics.Statistics.Moments

Documentation

data Moments

The first few central moments and covariances

Constructors

NoMoments

Moments

Fields

_rawCount :: !Int64

_rawMeans :: !(Vector Double)

_rawVariances :: !(Vector Double)

_rawCovariances :: !(Vector (Vector Double))

_rawSkewnesses :: !(Vector Double)

_rawKurtoses :: !(Vector Double)

class HasMoments t where

Methods

moments :: Lens' t Moments

rawCount :: Traversal' t Int64

rawCovariances :: Traversal' t (Vector (Vector Double))

rawKurtoses :: Traversal' t (Vector Double)

rawMeans :: Traversal' t (Vector Double)

rawSkewnesses :: Traversal' t (Vector Double)

rawVariances :: Traversal' t (Vector Double)

dim :: Int -> Moments

Generate an empty set of moments for a restricted number of dimensions

variances :: HasMoments t => IndexedTraversal' Int t Double

Extract the variance as a moment around the mean.

NB: setting this will render the higher Moments inconsistent.

variance :: HasMoments t => Int -> IndexedTraversal' Int t Double

stddevs :: HasMoments t => IndexedTraversal' Int t Double

NB: setting this will render the higher Moments inconsistent.

stddev :: HasMoments t => Int -> IndexedTraversal' Int t Double

Retrieve the nth std deviation from our set.

singleton :: Unbox a => a -> Vector a

momentsOf :: Foldable f => Getting Moments s (f Double) -> s -> Moments